backend/pyproject.toml
[project]
name = "fastapi-todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "bcrypt>=4.3.0",
    "beanie>=2.0.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.116.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
]
